fix(cookies): keep better-call out of the client bundle
