+++ packages/better-auth/src/cookies/cookie-utils.ts
@@ -1,41 +1,3 @@
-import { BetterAuthError } from "@better-auth/core/error";
-import type { CookieOptions } from "better-call";
-import { serializeCookie } from "better-call";
-
-/**
- * Per-cookie byte ceiling.
- * Safari's ~4093 floor is the lowest among browsers.
- *
- * @see https://datatracker.ietf.org/doc/html/rfc6265#section-6.1
- */
-export const MAX_COOKIE_SIZE = 4093;
-
-/**
- * Max chunks per cookie.
- * A larger value does not belong in a cookie.
- */
-export const MAX_COOKIE_CHUNKS = 100;
-
-/**
- * Largest value that keeps the serialized cookie within {@link MAX_COOKIE_SIZE}.
- * Overhead is measured with the real `serializeCookie` writer so it cannot
- * drift from the wire. Assumes a wire-safe (e.g. base64url) value.
- */
-export function getMaxCookieValueSize(
-	name: string,
-	options: CookieOptions,
-): number {
-	// serializeCookie mutates options (e.g. forces `secure`), so copy them.
-	const overhead = serializeCookie(name, "", { ...options }).length;
-	const size = MAX_COOKIE_SIZE - overhead;
-	if (size <= 0) {
-		throw new BetterAuthError(
-			`Cookie "${name}" attributes leave no room for a value.`,
-		);
-	}
-	return size;
-}
-
 function tryDecode(str: string): string {
 	if (str.indexOf("%") === -1) return str;
 	try {
